requirements.txt
boto3
python-dotenv
langchain
langchain-openai
langchain-google-genai
huggingface_hub
